services/ai-glm-service/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
websockets==15.0.1
httpx==0.27.0
python-dotenv==1.0.1
numpy==1.26.3
torch==2.1.2
torchaudio==2.1.2
transformers==4.36.2
accelerate==0.26.1
soundfile==0.13.1
librosa==0.11.0
scipy==1.15.3
numba==0.60.0
llvmlite==0.43.0
decorator==5.2.1
audioread==3.1.0
lazy_loader==0.4
msgpack==1.1.2
joblib==1.5.3
pooch==1.8.2
psutil==7.1.3
regex==2025.11.3
safetensors==0.7.0
tokenizers==0.15.2
huggingface-hub==0.31.4
filelock==3.20.1
packaging==25.0
pyyaml==6.0.3